+++ tests/mocks.py
@@ -89,14 +89,15 @@ EXAMPLE_DICT_POPULATED = {
     'struct_key': {
         'key1': u'val1',
         'key2': 2,
-        'key3': [1, 2, 3, None, True, False],
+        'key3': [1, 2, 3, 4.44, None, True, False],
         'key4': u'čđć',
         'key5': {
             'dict_key_1': u'1',
             'dict_key_2': 30,
-            'dict_key_3': [u'a', u'b', u'c', 3, {u'h': u'bar', u'g': [1, 2], u'j': [], u'l': True,
-                                                 u'm': False}, None],
+            'dict_key_3': [u'a', u'b', u'c', 7, {u'h': u'bar', u'g': [1, 2, 33.33], u'j': [],
+                                                 u'l': True, u'm': False}, None],
             'dict_key_4': None,
+            'dict_key_5': 55.55
         },
         'key6': None,
         'key7': [],
@@ -108,7 +109,8 @@ EXAMPLE_DICT_POPULATED = {
             }
         },
         'key9': True,
-        'key10': False
+        'key10': False,
+        'key11': 11.123
     },
     'timestamp_key': dt,
     'geo_point_key': GeoPoint(-20.2, +160.5),
@@ -197,14 +199,15 @@ EXAMPLE_PB_POPULATED.timestamp_key.FromDatetime(dt)
 EXAMPLE_PB_POPULATED.struct_key.update({
     'key1': u'val1',
     'key2': 2,
-    'key3': [1, 2, 3, None, True, False],
+    'key3': [1, 2, 3, 4.44, None, True, False],
     'key4': u'čđć',
     'key5': {
         'dict_key_1': u'1',
         'dict_key_2': 30,
-        'dict_key_3': [u'a', u'b', u'c', 3, {u'h': u'bar', u'g': [1, 2], u'j': [], u'l': True,
-                                             u'm': False}, None],
-        'dict_key_4': None
+        'dict_key_3': [u'a', u'b', u'c', 7, {u'h': u'bar', u'g': [1, 2, 33.33], u'j': [],
+                                             u'l': True, u'm': False}, None],
+        'dict_key_4': None,
+        'dict_key_5': 55.55
     },
     'key6': None,
     'key7': [],
@@ -216,7 +219,8 @@ EXAMPLE_PB_POPULATED.struct_key.update({
         }
     },
     'key9': True,
-    'key10': False
+    'key10': False,
+    'key11': 11.123
 })
 
 geo_point_value = latlng_pb2.LatLng(latitude=-20.2, longitude=+160.5)